+++ token-exchange/test/Coins.t.sol
@@ -19,7 +19,7 @@ contract CoinsTest is Test {
         assertEq(skillsCoin.owner(), address(this));
     }
 
-    function testMint(address _address,uint _amount) public {
+    function testMint(address _address, uint256 _amount) public {
         assertEq(skillsCoin.balanceOf(_address), 0);
         vm.prank(address(_address));
         skillsCoin.mint(_address, _amount);
@@ -30,27 +30,20 @@ contract CoinsTest is Test {
         assertEq(rareCoin.source(), address(skillsCoin));
     }
 
-    function testMintRareCoinFailed(address _address, uint _amount) public {
+    function testMintRareCoinFailed(address _address, uint256 _amount) public {
         vm.expectRevert();
-        (bool s,) = address(rareCoin).call(
-            abi.encodeWithSignature(
-                "mint(address,uint256)",
-                _address,
-                _amount
-            )
-        );
+        (bool s,) = address(rareCoin).call(abi.encodeWithSignature("mint(address,uint256)", _address, _amount));
         require(s, "Cannot mint");
     }
 
-    function testTrade(address _address, uint _amount) public {
+    function testTrade(address _address, uint256 _amount) public {
         vm.startPrank(_address);
         skillsCoin.mint(_address, _amount);
         assertEq(skillsCoin.balanceOf(_address), _amount);
 
         skillsCoin.approve(address(rareCoin), _amount);
-        uint allowance = skillsCoin.allowance(_address, address(rareCoin));
+        uint256 allowance = skillsCoin.allowance(_address, address(rareCoin));
         assertEq(allowance, _amount);
-
 
         rareCoin.trade(_amount);
         assertEq(skillsCoin.balanceOf(address(rareCoin)), _amount);
@@ -58,5 +51,4 @@ contract CoinsTest is Test {
         assertEq(rareCoin.balanceOf(_address), _amount);
         vm.stopPrank();
     }
-
 }
